+++ GEOSwiftTests/GEOS/GeometryConvertible+GEOSTests.swift
@@ -561,7 +561,8 @@ final class GeometryConvertible_GEOSTests: XCTestCase {
             Point(x: 1.5, y: 1),
             Point(x: 1, y: 1),
             Point(x: 1, y: 0)]))
-        XCTAssertEqual(try? poly.difference(with: unitPoly).isTopologicallyEquivalent(to: expectedPoly), true)
+        XCTAssertEqual(try? poly.difference(with: unitPoly)?.isTopologicallyEquivalent(to: expectedPoly),
+                       true)
     }
 
     func testDifferenceAllPairs() {
@@ -665,6 +666,21 @@ final class GeometryConvertible_GEOSTests: XCTestCase {
         }
     }
 
+    func testMinimumBoundingCircleLineString() {
+        XCTAssertEqual(try lineString1.minimumBoundingCircle(),
+                       Circle(center: Point(x: 0.5, y: 0), radius: 0.5))
+    }
+
+    func testMinimumBoundingCircleAllTypes() {
+        for g in geometryConvertibles {
+            do {
+                _ = try g.minimumBoundingCircle()
+            } catch {
+                XCTFail("Unexpected error for \(g) minimumBoundingCircle() \(error)")
+            }
+        }
+    }
+
     func testPolygonizeAllTypes() {
         for g in geometryConvertibles {
             do {
@@ -711,6 +727,16 @@ final class GeometryConvertible_GEOSTests: XCTestCase {
                 _ = try geometry.buffer(by: 0.5)
             } catch {
                 XCTFail("Unexpected error for \(geometry) buffer(by: 10) \(error)")
+            }
+        }
+    }
+
+    func testNegativeBufferWidthThrows() {
+        XCTAssertThrowsError(try Point(x: 0, y: 0).buffer(by: -1)) { (error) in
+            if case GEOSwiftError.negativeBufferWidth = error {
+                // pass
+            } else {
+                XCTFail("Threw unexpected error: \(error)")
             }
         }
     }